+++ test/test-expects.cake.dart
@@ -2,13 +2,6 @@ import 'package:cake/cake.dart';
 
 void main(List<String> arguments) async {
   TestRunnerDefault('Simple Test with no groups', [
-    // Generic Constructor
-    Test(
-      'True is true - assertion',
-      assertions: ((context) => [
-            Expect(ExpectType.equals, expected: true, actual: true),
-          ]),
-    ),
     Test(
       'True is true, set in setup',
       setup: (context) {